refactor: move apollo module into own file
